use same validation against death report on both study termination forms
